+++ aasm/generating/python_code.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import List
+from typing import List, Set
 
 
 class PythonCode:
@@ -8,6 +8,7 @@ class PythonCode:
         self.indent_size = indent_size
         self.indent: int = 0
         self.code_lines: List[str] = []
+        self.required_imports: Set[str] = set()
 
     def indent_left(self) -> None:
         self.indent -= self.indent_size
@@ -15,7 +16,9 @@ class PythonCode:
     def indent_right(self) -> None:
         self.indent += self.indent_size
 
-    def add_line(self, line: str) -> None:
+    def add_line(self, line: str, required_imports: Set[str] | None = None) -> None:
+        if required_imports is not None:
+            self.required_imports.update(required_imports)
         self.code_lines.append(self.indent * " " + line + "\n")
 
     def add_newline(self) -> None:
@@ -24,6 +27,13 @@ class PythonCode:
     def add_newlines(self, count: int) -> None:
         for _ in range(count):
             self.add_newline()
+
+    def add_required_imports(self) -> None:
+        lines: List[str] = []
+        for required_import in self.required_imports:
+            lines.append(f"import {required_import}\n")
+        lines.sort()
+        self.code_lines = lines + self.code_lines
 
     def add_template(self, template: str, **kwargs: Any) -> None:
         lines = template.render(kwargs).splitlines()